+++ src/app/core/models/entity.py
@@ -16,52 +16,6 @@ entity_schema = {
     "required": [ "entity_type_id", "content" ],
 }
 
-entity_type_schema = {
-    "type": "object",
-    "properties": {
-        "name": {"type": "string"},
-        "schema": {"type": "object"},
-    },
-    "required": [ "name", "schema" ],
-}
-
-class EntityType(db.Model):
-    """ **EntityType** db model """
-
-    __tablename__ = 'entity_type'
-
-    id = db.Column(
-        db.String(length=60),
-        unique=True,
-        nullable=False,
-        primary_key=True
-    )
-
-    name = db.Column(
-        db.String(length=80),
-        unique=True,
-        nullable=False,
-        default=''
-    )
-
-    schema = db.Column(
-        db.Text(),
-        nullable=False
-    )
-
-    def __init__(self, **kwargs):
-        self.id = uuid.uuid4().__str__()
-        self.name = kwargs.get('name')
-        self.schema = json.dumps(kwargs.get('schema'))
-
-    def toDict(self):
-        data = dict([])
-        data['id'] = self.id
-        data['name'] = self.name
-        data['schema'] = json.loads(self.schema)
-        return data
-
-
 class Entity(db.Model):
     """ **EntityType** db model """
